move explorer query logic to @sd/client
