+++ packages/client/src/explorer/usePathsExplorerQuery.ts
@@ -1,12 +1,13 @@
-import { FilePathOrder, FilePathSearchArgs, useLibraryQuery } from '@sd/client';
-
-import { UseExplorerSettings } from '../useExplorer';
+import { FilePathOrder, FilePathSearchArgs } from '../core';
+import { useLibraryQuery } from '../rspc';
 import { useExplorerQuery } from './useExplorerQuery';
 import { usePathsOffsetInfiniteQuery } from './usePathsOffsetInfiniteQuery';
 
 export function usePathsExplorerQuery(props: {
 	arg: FilePathSearchArgs;
-	explorerSettings: UseExplorerSettings<FilePathOrder>;
+	order: FilePathOrder | null;
+	/** This callback will fire any time the query successfully fetches new data. (NOTE: This will be removed on the next major version (react-query)) */
+	onSuccess?: () => void;
 }) {
 	const query = usePathsOffsetInfiniteQuery(props);
