+++ packages/client/src/index.ts
@@ -33,3 +33,4 @@ export * from './form';
 export * from './cache';
 export * from './color';
 export * from './solid';
+export * from './explorer';
